Reorganize code-utils to be more TypeScripty
