+++ packages/code-utils/package.json
@@ -5,13 +5,13 @@
   "author": "",
   "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/code-utils",
   "version": "1.2.22",
-  "main": "dist/index.js",
+  "main": "dist/src/index.js",
   "scripts": {
     "build": "tsc",
     "prepare": "yarn build",
     "test": "mocha --exit -r ts-node/register test/**/*.test.ts"
   },
-  "types": "./typings/index.d.ts",
+  "types": "./dist/src/index.d.ts",
   "dependencies": {
     "borc": "^2.1.2",
     "source-map-support": "^0.5.19"
