+++ packages/code-utils/tsconfig.json
@@ -25,7 +25,7 @@
     ]
   },
   "include": [
-    "./**/*.ts",
+    "./src/**/*.ts",
     "./typings/**/*.d.ts"
   ]
 }
